Updated render and bounding boxes to use floats.  Fixes issue with tiled, rotated, textures not lining up correctly.

+++ src/display/window.h
@@ -20,7 +20,7 @@ public:
   void Free(void);
   void Clear(void) const;
   void Show(void) const;
-  void View(int x, int y, float zoom);
+  void View(float x, float y, float zoom);
 private:
   std::shared_ptr<class WindowImpl> impl_;
 };
